requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.3.4
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
python-multipart==0.0.6
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
tenacity==8.2.3
python-dateutil==2.8.2
